@@ -128,12 +128,12 @@ class ChannelPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) => Navigator(
         onGenerateRoute: (settings) => MaterialPageRoute(
-          builder: (context) => Scaffold(
-            appBar: const StreamChannelHeader(
+          builder: (context) => const Scaffold(
+            appBar: StreamChannelHeader(
               showBackButton: false,
             ),
             body: Column(
-              children: const <Widget>[
+              children: <Widget>[
                 Expanded(
                   child: StreamMessageListView(),
                 ),
