@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/src/theme/avatar_theme.dart';
@@ -27,6 +29,7 @@ class StreamMessageThemeData with Diagnosticable {
     this.urlAttachmentTitleStyle,
     this.urlAttachmentTextStyle,
     this.urlAttachmentTitleMaxLine,
+    this.urlAttachmentTextMaxLine,
   }) : urlAttachmentBackgroundColor =
             urlAttachmentBackgroundColor ?? linkBackgroundColor;
 
@@ -82,6 +85,9 @@ class StreamMessageThemeData with Diagnosticable {
   /// Max number of lines in Url link title.
   final int? urlAttachmentTitleMaxLine;
 
+  /// Max number of lines in Url link text.
+  final int? urlAttachmentTextMaxLine;
+
   /// Copy with a theme
   StreamMessageThemeData copyWith({
     TextStyle? messageTextStyle,
@@ -102,6 +108,7 @@ class StreamMessageThemeData with Diagnosticable {
     TextStyle? urlAttachmentTitleStyle,
     TextStyle? urlAttachmentTextStyle,
     int? urlAttachmentTitleMaxLine,
+    int? urlAttachmentTextMaxLine,
   }) {
     return StreamMessageThemeData(
       messageTextStyle: messageTextStyle ?? this.messageTextStyle,
@@ -128,6 +135,8 @@ class StreamMessageThemeData with Diagnosticable {
           urlAttachmentTextStyle ?? this.urlAttachmentTextStyle,
       urlAttachmentTitleMaxLine:
           urlAttachmentTitleMaxLine ?? this.urlAttachmentTitleMaxLine,
+      urlAttachmentTextMaxLine:
+          urlAttachmentTextMaxLine ?? this.urlAttachmentTextMaxLine,
     );
   }
 
@@ -178,6 +187,16 @@ class StreamMessageThemeData with Diagnosticable {
         b.urlAttachmentTitleStyle,
         t,
       ),
+      urlAttachmentTitleMaxLine: lerpDouble(
+        a.urlAttachmentTitleMaxLine,
+        b.urlAttachmentTitleMaxLine,
+        t,
+      )?.round(),
+      urlAttachmentTextMaxLine: lerpDouble(
+        a.urlAttachmentTextMaxLine,
+        b.urlAttachmentTextMaxLine,
+        t,
+      )?.round(),
     );
   }
 
@@ -206,6 +225,7 @@ class StreamMessageThemeData with Diagnosticable {
       urlAttachmentTitleStyle: other.urlAttachmentTitleStyle,
       urlAttachmentTextStyle: other.urlAttachmentTextStyle,
       urlAttachmentTitleMaxLine: other.urlAttachmentTitleMaxLine,
+      urlAttachmentTextMaxLine: other.urlAttachmentTextMaxLine,
     );
   }
 
@@ -229,7 +249,8 @@ class StreamMessageThemeData with Diagnosticable {
           urlAttachmentHostStyle == other.urlAttachmentHostStyle &&
           urlAttachmentTitleStyle == other.urlAttachmentTitleStyle &&
           urlAttachmentTextStyle == other.urlAttachmentTextStyle &&
-          urlAttachmentTitleMaxLine == other.urlAttachmentTitleMaxLine;
+          urlAttachmentTitleMaxLine == other.urlAttachmentTitleMaxLine &&
+          urlAttachmentTextMaxLine == other.urlAttachmentTextMaxLine;
 
   @override
   int get hashCode =>
@@ -248,7 +269,8 @@ class StreamMessageThemeData with Diagnosticable {
       urlAttachmentHostStyle.hashCode ^
       urlAttachmentTitleStyle.hashCode ^
       urlAttachmentTextStyle.hashCode ^
-      urlAttachmentTitleMaxLine.hashCode;
+      urlAttachmentTitleMaxLine.hashCode ^
+      urlAttachmentTextMaxLine.hashCode;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -284,6 +306,10 @@ class StreamMessageThemeData with Diagnosticable {
       ..add(DiagnosticsProperty(
         'urlAttachmentTitleMaxLine',
         urlAttachmentTitleMaxLine,
+      ))
+      ..add(DiagnosticsProperty(
+        'urlAttachmentTextMaxLine',
+        urlAttachmentTextMaxLine,
       ));
   }
 }
