@@ -37,9 +37,8 @@ void main() {
     double? minCacheExtent,
     bool variableHeight = false,
   }) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     await tester.pumpWidget(
       MaterialApp(
@@ -105,9 +104,8 @@ void main() {
   testWidgets('List positioned with 0 at top - use default values',
       (WidgetTester tester) async {
     final itemPositionsListener = ItemPositionsListener.create();
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     await tester.pumpWidget(
       MaterialApp(
@@ -1543,9 +1541,8 @@ void main() {
     final itemPositionsListener = ItemPositionsListener.create();
     final itemScrollController = ItemScrollController();
 
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     await tester.pumpWidget(
       MaterialApp(
@@ -1600,9 +1597,8 @@ void main() {
     final itemPositionsListener = ItemPositionsListener.create();
     final itemScrollController = ItemScrollController();
 
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     await tester.pumpWidget(
       MaterialApp(
@@ -1661,9 +1657,8 @@ void main() {
       final itemPositionsListener = ItemPositionsListener.create();
       final itemScrollController = ItemScrollController();
 
-      tester.binding.window.devicePixelRatioTestValue = 1.0;
-      tester.binding.window.physicalSizeTestValue =
-          const Size(screenWidth, screenHeight);
+      tester.view.devicePixelRatio = 1.0;
+      tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
       await tester.pumpWidget(
         MaterialApp(
@@ -1732,9 +1727,8 @@ void main() {
 
   testWidgets('Jump to 100 then set itemCount to 0',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     final itemScrollController = ItemScrollController();
     final itemPositionsListener = ItemPositionsListener.create();
@@ -1781,9 +1775,8 @@ void main() {
 
   testWidgets('List positioned with 100 at top then set itemCount to 100',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     final itemCount = ValueNotifier<int>(defaultItemCount);
 
@@ -1823,9 +1816,8 @@ void main() {
 
   testWidgets('List positioned with 499 at bottom then set itemCount to 100',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     final itemCount = ValueNotifier<int>(defaultItemCount);
 
@@ -1943,9 +1935,8 @@ void main() {
   });
 
   testWidgets('Rebuild with scroll controller', (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
     final key = ValueNotifier<Key>(const ValueKey('key'));
     final itemScrollController = ItemScrollController();
 
@@ -1986,9 +1977,8 @@ void main() {
 
   testWidgets('Double rebuild with scroll controller',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
     final outerKey = ValueNotifier<Key>(const ValueKey('outerKey'));
     final innerKey = GlobalKey();
     final listKey = ValueNotifier<Key>(const ValueKey(null));
@@ -2036,9 +2026,8 @@ void main() {
   });
 
   testWidgets('Key change with scroll controller', (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
     final key = ValueNotifier<Key>(const ValueKey('key'));
     final itemScrollController = ItemScrollController();
 
@@ -2070,9 +2059,8 @@ void main() {
   });
 
   testWidgets('Scroll after rebuild', (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
     final key = ValueNotifier<Key>(const ValueKey('key'));
     final itemScrollController = ItemScrollController();
 
@@ -2113,9 +2101,8 @@ void main() {
 
   testWidgets('Scroll after rebuild when resusing state',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
     final containerKey = ValueNotifier<Key>(const ValueKey('key'));
     final scrollKey = GlobalKey();
     final itemScrollController = ItemScrollController();
@@ -2158,9 +2145,8 @@ void main() {
 
   testWidgets('Scroll after changing scroll controller',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     final itemScrollController0 = ItemScrollController();
     final itemScrollController1 = ItemScrollController();
@@ -2205,9 +2191,8 @@ void main() {
 
   testWidgets('Scroll after swapping scroll controllers',
       (WidgetTester tester) async {
-    tester.binding.window.devicePixelRatioTestValue = 1.0;
-    tester.binding.window.physicalSizeTestValue =
-        const Size(screenWidth, screenHeight);
+    tester.view.devicePixelRatio = 1.0;
+    tester.view.physicalSize = const Size(screenWidth, screenHeight);
 
     final itemScrollController0 = ItemScrollController();
     final itemScrollController1 = ItemScrollController();
