@@ -40,10 +40,12 @@ void main() {
   });
 
   setUp(() {
-    methodChannel.setMockMethodCallHandler((MethodCall methodCall) async {
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(methodChannel, (MethodCall methodCall) async {
       if (methodCall.method == 'listen') {
         try {
-          await ServicesBinding.instance.defaultBinaryMessenger
+          await TestDefaultBinaryMessengerBinding
+              .instance.defaultBinaryMessenger
               .handlePlatformMessage(
             methodChannel.name,
             methodChannel.codec.encodeSuccessEnvelope('wifi'),
@@ -53,6 +55,7 @@ void main() {
           print(e);
         }
       }
+      return null;
     });
   });
 
@@ -106,6 +109,7 @@ void main() {
   });
 
   tearDown(() {
-    methodChannel.setMockMethodCallHandler(null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(methodChannel, null);
   });
 }
