@@ -69,11 +69,11 @@ final _messageInputThemeControlMidLerp = StreamMessageInputThemeData(
   borderRadius: BorderRadius.circular(20),
   sendAnimationDuration: const Duration(milliseconds: 300),
   inputBackgroundColor: const Color(0xff87898b),
-  actionButtonColor: const Color(0xff005fff),
+  actionButtonColor: const Color(0xff196eff),
   actionButtonIdleColor: const Color(0xff7a7a7a),
-  sendButtonColor: const Color(0xff005fff),
+  sendButtonColor: const Color(0xff196eff),
   sendButtonIdleColor: const Color(0xff848585),
-  expandButtonColor: const Color(0xff005fff),
+  expandButtonColor: const Color(0xff196eff),
   inputTextStyle: const TextStyle(
     color: Color(0xff7f7f7f),
     fontSize: 14,
