@@ -15,10 +15,10 @@ class Users extends Table {
   TextColumn get language => text().nullable()();
 
   /// Date of user creation
-  DateTimeColumn get createdAt => dateTime().withDefault(currentDateAndTime)();
+  DateTimeColumn get createdAt => dateTime().nullable()();
 
   /// Date of last user update
-  DateTimeColumn get updatedAt => dateTime().withDefault(currentDateAndTime)();
+  DateTimeColumn get updatedAt => dateTime().nullable()();
 
   /// Date of last user connection
   DateTimeColumn get lastActive => dateTime().nullable()();
