@@ -96,11 +96,11 @@ void main() {
     expect(message.updatedAt, isSameDateAs(entity.updatedAt));
     expect(message.extraData, entity.extraData);
     expect(message.user!.id, entity.userId);
-    expect(message.deletedAt, isSameDateAs(entity.deletedAt!));
+    expect(message.deletedAt, isSameDateAs(entity.deletedAt));
     expect(message.text, entity.messageText);
     expect(message.pinned, entity.pinned);
-    expect(message.pinExpires, isSameDateAs(entity.pinExpires!));
-    expect(message.pinnedAt, isSameDateAs(entity.pinnedAt!));
+    expect(message.pinExpires, isSameDateAs(entity.pinExpires));
+    expect(message.pinnedAt, isSameDateAs(entity.pinnedAt));
     expect(message.pinnedBy!.id, entity.pinnedByUserId);
     expect(message.reactionCounts, entity.reactionCounts);
     expect(message.reactionScores, entity.reactionScores);
@@ -191,11 +191,11 @@ void main() {
     expect(entity.updatedAt, isSameDateAs(message.updatedAt));
     expect(entity.extraData, message.extraData);
     expect(entity.userId, message.user!.id);
-    expect(entity.deletedAt, isSameDateAs(message.deletedAt!));
+    expect(entity.deletedAt, isSameDateAs(message.deletedAt));
     expect(entity.messageText, message.text);
     expect(entity.pinned, message.pinned);
-    expect(entity.pinExpires, isSameDateAs(message.pinExpires!));
-    expect(entity.pinnedAt, isSameDateAs(message.pinnedAt!));
+    expect(entity.pinExpires, isSameDateAs(message.pinExpires));
+    expect(entity.pinnedAt, isSameDateAs(message.pinnedAt));
     expect(entity.pinnedByUserId, message.pinnedBy!.id);
     expect(entity.reactionCounts, message.reactionCounts);
     expect(entity.reactionScores, message.reactionScores);
